pubsub/py/pyproject.toml
[tool.poetry]
name = "davidkhala.gcp.pubsub"
version = "0.0.0"
description = "@davidkhala/gcp/pubsub"
authors = ["David Liu <david-khala@hotmail.com>"]
readme = "README.md"
packages = [{ include = "davidkhala" }]

[tool.poetry.dependencies]
python = "^3.10"
davidkhala-gcp = "*"
google-cloud-pubsub = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
